// features/api/model/productResource.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * Petrol Integrity System API Documentation
 * OpenAPI spec version: 0.0.1
 */

export interface ProductResource {
  id: string;
  name: string;
  current_price: number;
  vat_rate: number;
  /** Default to 16% if null */
  organization_id: string;
  created_at: string;
  updated_at: string;
}