// features/api/model/storeProductRequest.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * Petrol Integrity System API Documentation
 * OpenAPI spec version: 0.0.1
 */

export interface StoreProductRequest {
  /** @maxLength 255 */
  name: string;
  /** @minimum 0 */
  current_price: number;
  /**
   * @minimum 0
   * @maximum 100
   * @nullable
   */
  vat_rate?: number | null;
}